removing old training script/updating to new lightning setup
